core: Fix CID 292193 argument cannot be negative
Calling socket() can fail. In this case socket() returns -1. In this
case it does not make sense to continue and use a broken listenfd when
calling bind().

+++ core/network_thread.c
@@ -123,6 +123,9 @@ int listener_create(const char *path, int type)
 	if (listenfd == -1) {
 		TRACE("creating socket at %s", path);
 		listenfd = socket(AF_LOCAL, type, 0);
+		if (listenfd < 0) {
+			return -1;
+		}
 		unlink(path);
 		bzero(&servaddr, sizeof(servaddr));
 		servaddr.sun_family = AF_LOCAL;
@@ -142,7 +145,6 @@ int listener_create(const char *path, int type)
 			return -1;
 		}
 	return listenfd;
-
 }
 
 static void cleanum_msg_list(void)
